requirements.txt
python-binance
python-dotenv
loguru
